+++ src/utils.ts
@@ -70,6 +70,10 @@ export const isSoftphoneJid = function (jid: string): boolean {
   return !!jid.match(/.*@.*gjoll.*/i);
 };
 
+export const isPeerVideoJid = function (jid: string) {
+  return isVideoJid(jid) && jid.startsWith('peer-');
+};
+
 export const isVideoJid = function (jid: string): boolean {
   return jid && !!jid.match(/@conference/) && !isAcdJid(jid);
 };