+++ test/test-pages/common/sdk-controller.js
@@ -4,7 +4,7 @@ import { getSdk, PureCloudWebrtcSdk } from '../sdk-proxy';
 import utils from './utils';
 
 let currentSession;
-let startVideoOpts;
+let videoOpts;
 let currentSessionId;
 let webrtcSdk;
 let conversationsApi;
@@ -78,8 +78,8 @@ async function endSession () {
     const controls = document.getElementById('video-actions');
     controls.classList.add('hidden');
 
-    const startControls = document.getElementById('start-controls');
-    startControls.classList.remove('hidden');
+    const startControls = document.querySelectorAll('.start-controls');
+    startControls.forEach(el => el.classList.remove('hidden'));
 
     utils.writeToLog('Call ended');
   } catch (err) {
@@ -123,6 +123,7 @@ function ready () {
 function pendingSession (options) {
   let output = `${_getLogHeader('pendingSession')}
     id: ${JSON.stringify(options.id)}
+    sessionType: ${JSON.stringify(options.sessionType)}
     address: ${JSON.stringify(options.address)}
     conversationId: ${JSON.stringify(options.conversationId)}
     autoAnswer: ${JSON.stringify(options.autoAnswer)}`;
@@ -175,19 +176,19 @@ async function sessionStarted (session) {
 
     let mediaStream;
 
-    if (!startVideoOpts.video && !startVideoOpts.audio) {
+    if (!videoOpts.video && !videoOpts.audio) {
       mediaStream = new MediaStream();
-    } else if (!startVideoOpts.video || !startVideoOpts.audio || startVideoOpts.videoResolution) {
-      if (startVideoOpts.video) {
-        startVideoOpts.video = getDeviceId('video');
+    } else if (!videoOpts.video || !videoOpts.audio || videoOpts.videoResolution) {
+      if (videoOpts.video) {
+        videoOpts.video = getDeviceId('video');
       }
 
-      if (startVideoOpts.audio) {
-        startVideoOpts.audio = getDeviceId('audio');
+      if (videoOpts.audio) {
+        videoOpts.audio = getDeviceId('audio');
       }
 
-      console.log({ startVideoOpts });
-      mediaStream = await webrtcSdk.createMedia(startVideoOpts);
+      console.log({ videoOpts });
+      mediaStream = await webrtcSdk.createMedia(videoOpts);
     }
 
     const sessionEventsToLog = ['participantsUpdate', 'activeVideoParticipantsUpdate', 'speakersUpdate'];
@@ -331,14 +332,7 @@ function connected (e) {
   utils.writeToLog('connected event', e);
 }
 
-async function startVideoConference ({ noAudio, noVideo, mediaStream, useConstraints } = {}) {
-  const roomJid = getInputValue('video-jid');
-  if (!roomJid) {
-    const message = 'roomJid required to start a video call';
-    document.getElementById('log-data').value += `${message}\n`;
-    throw new Error(message);
-  }
-
+async function startVideoConference ({ noAudio, noVideo, mediaStream, useConstraints } = {}, answerPendingSession) {
   let videoResolution;
 
   if (useConstraints) {
@@ -346,14 +340,26 @@ async function startVideoConference ({ noAudio, noVideo, mediaStream, useConstra
     console.log('proceeding with custom resolution', videoResolution);
   }
 
-  startVideoOpts = { video: !noVideo, audio: !noAudio, mediaStream, videoResolution };
+  videoOpts = { video: !noVideo, audio: !noAudio, mediaStream, videoResolution };
 
-  webrtcSdk.startVideoConference(roomJid);
+  if (answerPendingSession) {
+    webrtcSdk.acceptPendingSession(currentSessionId);
+  } else {
+    const roomJid = getInputValue('video-jid');
+    if (!roomJid) {
+      const message = 'roomJid required to start a video call';
+      document.getElementById('log-data').value += `${message}\n`;
+      throw new Error(message);
+    }
+
+    webrtcSdk.startVideoConference(roomJid, getInputValue('invitee-jid'));
+  }
+
   const element = document.getElementById('waiting-for-media');
   element.classList.remove('hidden');
 
-  const startControls = document.getElementById('start-controls');
-  startControls.classList.add('hidden');
+  const startControls = document.querySelectorAll('.start-controls');
+  startControls.forEach(el => el.classList.add('hidden'));
 }
 
 function setVideoMute (mute) {